+++ python/NEWS.md
@@ -2,7 +2,7 @@
 
 ## New Features
 - Some datasets now have "extras" stored as JSON, which will be downloaded and used to provide extra data like data profiles, bibtex, dataset licensing info, etc.
-- `OpenTabularDataset` has new properties: `has_extras`, `license`, `bibtex` (not all datasets have meaningful values yet)
+- `OpenTabularDataset` has new properties: `has_extras`, `license`, `bibtex`, `categorical_attributes` (not all datasets have meaningful values yet); and a new method: `has_extra`
 - new datasets: 
   - `duolingo-original`
   - `duolingo-categorical`